Improve organization of files and components, because this will get complicated, fast.
